requirements.txt
backcall==0.1.0
decorator==4.4.1
dj-database-url==0.5.0
Django==2.2.7
django-extensions==2.2.6
django-filter==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
PyJWT==1.7.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.8
whitenoise==4.1.4
